dleeted unused notebooks and cleaned up substantially.
